+++ apx_sdk/CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed error message in case of invalid `verificationMethod`.
 - Changed internal value of `HttpSignatureError::UrlError` to string.
 - Allow URLs in `create_http_signature_rfc9421`.
+- Always add digest header during signing if request body is provided.
 
 ### Fixed